+++ src/antlr/v4/runtime/atn/LexerTypeAction.d
@@ -78,9 +78,9 @@ class LexerTypeAction : LexerAction
      * @safe
      * @nothrow
      */
-    public int hashCode() @safe nothrow
+    public size_t hashCode() @safe nothrow
     {
-        int hash = MurmurHash.initialize();
+        size_t hash = MurmurHash.initialize();
         hash = MurmurHash.update(hash, Utils.rank(getActionType));
         hash = MurmurHash.update(hash, type);
         return MurmurHash.finish(hash, 2);
